function_lda_bigquery/requirements.txt
# Function dependencies, for example:
# package>=version
gensim
setuptools
wheel
spacy
langdetect
textblob
nltk
google-cloud-storage